[DOK-225] refactor : 폴더구조 리팩토링 ( 상품 등록 / 수정 )에서 오류 발생 )
