+++ src/Router.js
@@ -1,44 +1,45 @@
-import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import MainPage from "pages/User/MainPage/MainPage";
-import ProductDetail from "pages/User/ProductDetailPage/ProductDetailPage";
-import Cart from "pages/User/CartPage/CartPage";
-import OrderList from "pages/User/MyPage/Order/OrderListPage/OrderListPage";
-import SignUp from "pages/commonPage/SignUp";
+import { Payment } from "@mui/icons-material";
 import AdminListPage from "pages/Admin/AdminManage/AdminListPage";
 import AdminRegisterPage from "pages/Admin/AdminManage/AdminRegisterPage";
 import AnnouncementPage from "pages/Admin/CSManage/AnnouncementPage";
 import BannerManagementPage from "pages/Admin/CSManage/BannerManagementPage";
 import FAQPage from "pages/Admin/CSManage/FAQPage";
-import InventoryManagePage from "pages/Admin/ProductManage/InventoryManagePage";
-import MileageApprovalPage from "pages/Admin/UserManage/MileageApprovalPage";
-import MileageHistoryPage from "pages/Admin/UserManage/MileageHistoryPage";
 import MileageInfoPage from "pages/Admin/CSManage/MileageInfoPage";
+import ShippingInfoPage from "pages/Admin/CSManage/ShippingInfoPage";
 import OrderInquiryPage from "pages/Admin/InquiryManage/OrderInquiryPage";
-import OrderListPage from "pages/Admin/OrderManage/OrderListPage";
 import PaymentInquiryPage from "pages/Admin/InquiryManage/PaymentInquiryPage";
-import ProductEditPage from "pages/Admin/ProductManage/ProductEditPage";
 import ProductInquiryPage from "pages/Admin/InquiryManage/ProductInquiryPage";
+import OrderListPage from "pages/Admin/OrderManage/OrderListPage";
+import InventoryManagePage from "pages/Admin/ProductManage/InventoryManagePage";
+import ProductEditPage from "pages/Admin/ProductManage/ProductEditPage";
 import ProductListPage from "pages/Admin/ProductManage/ProductListPage";
 import ProductRegisterPage from "pages/Admin/ProductManage/ProductRegisterPage";
-import ShippingInfoPage from "pages/Admin/CSManage/ShippingInfoPage";
+import MileageApprovalPage from "pages/Admin/UserManage/MileageApprovalPage";
+import MileageHistoryPage from "pages/Admin/UserManage/MileageHistoryPage";
 import UserListPage from "pages/Admin/UserManage/UserListPage";
 import UserRegisterPage from "pages/Admin/UserManage/UserRegisterPage";
-import Login from "pages/commonPage/Login";
-import ManualPage from "pages/User/CustomerServicePage/ManualPage";
-import NoticePage from "pages/User/CustomerServicePage/NoticePage";
-import UserFAQPage from "pages/User/CustomerServicePage/UserFAQPage";
-import Refund from "pages/User/MyPage/Order/RefundPage/RefundPage";
-import Mileage from "pages/User/MyPage/MileagePage/MileagePage";
-import Inquiry from "pages/User/MyPage/Inquiry/InquiryWritePage/InquiryWritePage";
-import InquiryHistory from "pages/User/MyPage/Inquiry/InquiryHistoryPage";
-import Review from "pages/User/MyPage/Review/ReviewPage/ReviewPage";
-import MyInfo from "pages/User/MyPage/MyInfoPage/MyInfoPage";
-import Payment from "pages/User/PaymentPage/PaymentPage";
-import OrderDetailPage from "pages/User/MyPage/Order/OrderListPage/OrderDetailPage";
-import ReviewWrite from "pages/User/MyPage/Review/ReviewWritePage/ReviewWritePage";
-import ReviewEdit from "pages/User/MyPage/Review/ReviewWritePage/ReviewEditPage";
+import Cart from "pages/User/CartPage/CartPage";
 import CategoryPage from "pages/User/CategoryPage/CategoryPage";
+import { ManualPage } from "pages/User/CustomerServicePage/ManualPage";
+import { NoticePage } from "pages/User/CustomerServicePage/NoticePage";
+import { UserFAQPage } from "pages/User/CustomerServicePage/UserFAQPage";
+import MainPage from "pages/User/MainPage/MainPage";
+import InquiryHistory from "pages/User/MyPage/Inquiry/InquiryHistoryPage/InquiryHistoryPage";
+import Inquiry from "pages/User/MyPage/Inquiry/InquiryWritePage/InquiryWritePage";
+import Mileage from "pages/User/MyPage/MileagePage/MileagePage";
+import MyInfo from "pages/User/MyPage/MyInfoPage/MyInfoPage";
+import OrderDetailPage from "pages/User/MyPage/Order/OrderListPage/OrderDetailPage";
+import OrderList from "pages/User/MyPage/Order/OrderListPage/OrderListPage";
+import Refund from "pages/User/MyPage/Order/RefundPage/RefundPage";
+import Review from "pages/User/MyPage/Review/ReviewPage/ReviewPage";
+import ReviewEdit from "pages/User/MyPage/Review/ReviewWritePage/ReviewEditPage";
+import ReviewWrite from "pages/User/MyPage/Review/ReviewWritePage/ReviewWritePage";
+import ProductDetail from "pages/User/ProductDetailPage/ProductDetailPage";
+import SignUp from "pages/commonPage/SignUp";
+import React from "react";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { Login } from "tabler-icons-react";
+
 
 const Router = () => {
   return (
