+++ src/pages/Admin/AdminManage/AdminRegisterPage.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import AdminBar from "components/AdminBar";
+
 import {
   Paper,
   Box,
@@ -9,9 +9,11 @@ import {
   AlertTitle,
   Snackbar,
 } from "@mui/material";
-import { AdminButton, InputBoxS } from "components/AdminComponents";
 import IconButton from "@mui/joy/IconButton";
 import CloseRoundedIcon from "@mui/icons-material/CloseRounded";
+import AdminBar from "components/organisms/AdminBar";
+import { InputBoxS } from "components/atoms/Input";
+import { AdminButton } from "components/atoms/AdminCommonButton";
 
 const AdminRegisterPage = () => {
   // Declare selectedMenu and setSelectedMenu using useState
