+++ src/pages/Admin/CSManage/AnnouncementPage.js
@@ -1,11 +1,5 @@
 import React, { useEffect, useState } from "react";
 import styled from "styled-components";
-import {
-  AdminButton,
-  InputBoxS,
-  InputBoxM,
-  CustomSelect,
-} from "../../../components/atoms/CKEditor";
 import AdminBar from "../../../components/organisms/AdminBar";
 import SearchIcon from "@mui/icons-material/Search";
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
@@ -28,6 +22,10 @@ import {
   Toolbar,
   Typography,
 } from "@mui/material";
+import { InputBoxS } from "components/atoms/Input";
+import { AdminButton } from "components/atoms/AdminCommonButton";
+import { InputBoxM } from "components/atoms/Input";
+import { CustomSelect } from "components/atoms/AdminSelectBox";
 
 const dataList = [
   {
