+++ src/pages/Admin/ProductManage/InventoryManagePage.js
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from "react";
 import styled from "styled-components";
-import AdminBar from "components/AdminBar";
-import { InputBoxS, AdminButton } from "components/AdminComponents";
+import AdminBar from "components/organisms/AdminBar";
+import { InputBoxS } from "components/atoms/Input";
+import { AdminButton } from "components/atoms/AdminCommonButton";
 import SearchIcon from "@mui/icons-material/Search";
 import Input from "@mui/material/Input";
 import {
