+++ src/pages/Admin/UserManage/MileageHistoryPage.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import AdminBar from "components/AdminBar";
+import AdminBar from "components/organisms/AdminBar";
 import {
   Paper,
   Box,
@@ -13,9 +13,9 @@ import {
   Button,
 } from "@mui/material";
 import styled from "styled-components";
-import { InputBoxS } from "components/AdminComponents";
 import SearchIcon from "@mui/icons-material/Search";
 import CheckIcon from "@mui/icons-material/Check";
+import { InputBoxS } from "components/atoms/Input";
 
 const dataList = [
   {
