+++ src/pages/Admin/UserManage/UserListPage.js
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from "react";
 import styled from "styled-components";
-import AdminBar from "components/AdminBar";
+import AdminBar from "components/organisms/AdminBar";
 import SearchIcon from "@mui/icons-material/Search";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { InputBoxS, AdminButton } from "components/AdminComponents";
+import { InputBoxS } from "components/atoms/Input";
+import { AdminButton } from "components/atoms/AdminCommonButton";
 import { Box } from "@mui/system";
 import {
   Divider,
