+++ src/pages/User/CustomerServicePage/UserFAQPage.js
@@ -1,5 +1,5 @@
-import { CustomerServiceLayout } from "components/layout/CustomerServiceLayout";
-import { FAQBody } from "components/layout/body/CustomerServicePage/FAQBody";
+import { CustomerServiceLayout } from "components/templete/CustomerServiceLayout";
+import { FAQBody } from "./FAQBody";
 
 export const UserFAQPage = () => {
   return (
