+++ src/pages/User/MyPage/Inquiry/InquiryWritePage/InquiryWritePage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SidebarLayout from "../../../components/templates/SidebarLayout";
+import SidebarLayout from "components/templete/SidebarLayout";
 import InquiryWriteBody from "./InquiryWriteBody";
