+++ src/pages/User/MyPage/Order/OrderListPage/OrderDetailPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SidebarLayout from "components/layout/SidebarLayout";
+import SidebarLayout from "components/templete/SidebarLayout";
 import OrderDetailBody from "./OrderDetailBody";
 
 const OrderDetailPage = () => {
