+++ src/pages/commonPage/SignUp.js
@@ -4,10 +4,10 @@ import styled from "styled-components";
 import { colors } from "styles/commonTheme";
 import signUpImage from "assets/images/signUpPage.png";
 import "assets/font/font.css";
-import FloatingLabelInput from "components/FloatingLabelInput";
-import AddressField from "components/AddressField";
-import BasicDatePicker from "components/BasicDatePicker";
-import StyledButton from "components/StyledButton";
+import FloatingLabelInput from "components/molecules/FloatingLabelInput";
+import AddressField from "components/molecules/AddressField";
+import BasicDatePicker from "components/atoms/BasicDatePicker";
+import StyledButton from "components/Button&Modal";
 
 const Base = styled.div`
   width: 1920px;
